+++ src/char_sequences/wavelet_matrix.rs
@@ -33,16 +33,16 @@ use crate::Serializable;
 ///
 /// // It accepts an integer representable in 8 bits.
 /// let mut seq = CompactVector::new(8)?;
-/// seq.extend(text.chars().map(|c| c as usize))?;
+/// seq.extend(text.chars().map(|c| c as u64 ))?;
 /// let wm = WaveletMatrix::<Rank9Sel>::new(seq)?;
 ///
 /// assert_eq!(wm.len(), len);
-/// assert_eq!(wm.alph_size(), 'n' as usize + 1);
+/// assert_eq!(wm.alph_size(), 'n' as u64 + 1);
 /// assert_eq!(wm.alph_width(), 7);
 ///
-/// assert_eq!(wm.access(2), Some('n' as usize ));
-/// assert_eq!(wm.rank(3, 'a' as usize), Some(1));
-/// assert_eq!(wm.select(1, 'n' as usize), Some(4));
+/// assert_eq!(wm.access(2), Some('n' as u64));
+/// assert_eq!(wm.rank(3, 'a' as u64), Some(1));
+/// assert_eq!(wm.select(1, 'n' as u64), Some(4));
 /// # Ok(())
 /// # }
 /// ```
@@ -78,7 +78,7 @@ where
         }
 
         let alph_size = seq.iter().max().unwrap() + 1;
-        let alph_width = utils::needed_bits(alph_size);
+        let alph_width = utils::needed_bits(alph_size as u64);
 
         let mut zeros = seq;
         let mut ones = CompactVector::new(alph_width).unwrap();
@@ -147,11 +147,11 @@ where
     /// use sucds::int_vectors::CompactVector;
     ///
     /// let mut seq = CompactVector::new(8)?;
-    /// seq.extend("banana".chars().map(|c| c as usize))?;
+    /// seq.extend("banana".chars().map(|c| c as u64))?;
     /// let wm = WaveletMatrix::<Rank9Sel>::new(seq)?;
     ///
-    /// assert_eq!(wm.access(2), Some('n' as usize));
-    /// assert_eq!(wm.access(5), Some('a' as usize));
+    /// assert_eq!(wm.access(2), Some('n' as u64));
+    /// assert_eq!(wm.access(5), Some('a' as u64));
     /// assert_eq!(wm.access(6), None);
     /// # Ok(())
     /// # }
@@ -197,12 +197,12 @@ where
     /// use sucds::int_vectors::CompactVector;
     ///
     /// let mut seq = CompactVector::new(8)?;
-    /// seq.extend("banana".chars().map(|c| c as usize))?;
+    /// seq.extend("banana".chars().map(|c| c as u64))?;
     /// let wm = WaveletMatrix::<Rank9Sel>::new(seq)?;
     ///
-    /// assert_eq!(wm.rank(3, 'a' as usize), Some(1));
-    /// assert_eq!(wm.rank(5, 'c' as usize), Some(0));
-    /// assert_eq!(wm.rank(7, 'b' as usize), None);
+    /// assert_eq!(wm.rank(3, 'a' as u64), Some(1));
+    /// assert_eq!(wm.rank(5, 'c' as u64), Some(0));
+    /// assert_eq!(wm.rank(7, 'b' as u64), None);
     /// # Ok(())
     /// # }
     /// ```
@@ -232,12 +232,12 @@ where
     /// use sucds::int_vectors::CompactVector;
     ///
     /// let mut seq = CompactVector::new(8)?;
-    /// seq.extend("banana".chars().map(|c| c as usize))?;
+    /// seq.extend("banana".chars().map(|c| c as u64))?;
     /// let wm = WaveletMatrix::<Rank9Sel>::new(seq)?;
     ///
-    /// assert_eq!(wm.rank_range(1..4, 'a' as usize), Some(2));
-    /// assert_eq!(wm.rank_range(2..4, 'c' as usize), Some(0));
-    /// assert_eq!(wm.rank_range(4..7, 'b' as usize), None);
+    /// assert_eq!(wm.rank_range(1..4, 'a' as u64), Some(2));
+    /// assert_eq!(wm.rank_range(2..4, 'c' as u64), Some(0));
+    /// assert_eq!(wm.rank_range(4..7, 'b' as u64), None);
     /// # Ok(())
     /// # }
     /// ```
@@ -288,11 +288,11 @@ where
     /// use sucds::int_vectors::CompactVector;
     ///
     /// let mut seq = CompactVector::new(8)?;
-    /// seq.extend("banana".chars().map(|c| c as usize))?;
+    /// seq.extend("banana".chars().map(|c| c as u64))?;
     /// let wm = WaveletMatrix::<Rank9Sel>::new(seq)?;
     ///
-    /// assert_eq!(wm.select(1, 'a' as usize), Some(3));
-    /// assert_eq!(wm.select(0, 'c' as usize), None);
+    /// assert_eq!(wm.select(1, 'a' as u64), Some(3));
+    /// assert_eq!(wm.select(0, 'c' as u64), None);
     /// # Ok(())
     /// # }
     /// ```
@@ -349,7 +349,7 @@ where
     /// use sucds::int_vectors::CompactVector;
     ///
     /// let mut seq = CompactVector::new(8)?;
-    /// seq.extend("banana".chars().map(|c| c as usize))?;
+    /// seq.extend("banana".chars().map(|c| c as u64))?;
     /// let wm = WaveletMatrix::<Rank9Sel>::new(seq)?;
     ///
     /// assert_eq!(wm.quantile(1..4, 0), Some('a' as usize)); // The 0th in "ana" should be "a"
@@ -417,21 +417,21 @@ where
     /// use sucds::int_vectors::CompactVector;
     ///
     /// let mut seq = CompactVector::new(8)?;
-    /// seq.extend("banana".chars().map(|c| c as usize))?;
+    /// seq.extend("banana".chars().map(|c| c as u64))?;
     /// let wm = WaveletMatrix::<Rank9Sel>::new(seq)?;
     ///
     /// // Intersections among "ana", "na", and "ba".
     /// assert_eq!(
     ///     wm.intersect(&[1..4, 4..6, 0..2], 0),
-    ///     Some(vec!['a' as usize, 'b' as usize, 'n' as usize])
+    ///     Some(vec!['a' as u64, 'b' as u64, 'n' as u64])
     /// );
     /// assert_eq!(
     ///     wm.intersect(&[1..4, 4..6, 0..2], 1),
-    ///     Some(vec!['a' as usize, 'n' as usize])
+    ///     Some(vec!['a' as u64, 'n' as u64])
     /// );
     /// assert_eq!(
     ///     wm.intersect(&[1..4, 4..6, 0..2], 2),
-    ///     Some(vec!['a' as usize])
+    ///     Some(vec!['a' as u64])
     /// );
     /// assert_eq!(
     ///     wm.intersect(&[1..4, 4..6, 0..2], 3),
@@ -523,13 +523,13 @@ where
     /// use sucds::int_vectors::CompactVector;
     ///
     /// let mut seq = CompactVector::new(8)?;
-    /// seq.extend("ban".chars().map(|c| c as usize))?;
+    /// seq.extend("ban".chars().map(|c| c as u64))?;
     /// let wm = WaveletMatrix::<Rank9Sel>::new(seq)?;
     ///
     /// let mut it = wm.iter();
-    /// assert_eq!(it.next(), Some('b' as usize));
-    /// assert_eq!(it.next(), Some('a' as usize));
-    /// assert_eq!(it.next(), Some('n' as usize));
+    /// assert_eq!(it.next(), Some('b' as u64));
+    /// assert_eq!(it.next(), Some('a' as u64));
+    /// assert_eq!(it.next(), Some('n' as u64));
     /// assert_eq!(it.next(), None);
     /// # Ok(())
     /// # }
@@ -643,11 +643,11 @@ mod test {
         let len = text.chars().count();
 
         let mut seq = CompactVector::new(8).unwrap();
-        seq.extend(text.chars().map(|c| c as usize)).unwrap();
+        seq.extend(text.chars().map(|c| c as u64)).unwrap();
         let wm = WaveletMatrix::<Rank9Sel>::new(seq).unwrap();
 
         assert_eq!(wm.len(), len);
-        assert_eq!(wm.alph_size(), ('u' as usize) + 1);
+        assert_eq!(wm.alph_size(), ('u' as u64) + 1);
         assert_eq!(wm.alph_width(), 7);
 
         assert_eq!(wm.access(20), Some('h' as usize));
@@ -667,7 +667,7 @@ mod test {
     fn test_serialize() {
         let text = "tobeornottobethatisthequestion";
         let mut seq = CompactVector::new(8).unwrap();
-        seq.extend(text.chars().map(|c| c as usize)).unwrap();
+        seq.extend(text.chars().map(|c| c as u64)).unwrap();
         let wm = WaveletMatrix::<Rank9Sel>::new(seq).unwrap();
 
         let mut bytes = vec![];
